+++ src/world.jl
@@ -181,6 +181,94 @@ function _move_entity!(world::World, entity::Entity, archetype_index::UInt32)::I
     return new_row
 end
 
+function _copy_entity!(world::World, entity::Entity)::Entity
+    _check_locked(world)
+
+    index = world._entities[entity._id]
+    new_entity, new_row = _create_entity!(world, index.archetype)
+    archetype = world._archetypes[index.archetype]
+
+    for comp in archetype.components
+        _copy_component_data!(world, comp, index.archetype, index.archetype, index.row, UInt32(new_row))
+    end
+
+    world._entities[new_entity._id] = _EntityIndex(index.archetype, UInt32(new_row))
+
+    if _has_observers(world._event_manager, OnCreateEntity)
+        _fire_create_entity(world._event_manager, new_entity, archetype.mask)
+    end
+    return new_entity
+end
+
+@generated function _copy_entity!(
+    world::W,
+    entity::Entity,
+    ::Val{ATS},
+    add::Tuple,
+    ::RTS,
+)::Entity where {W<:World,ATS<:Tuple,RTS<:Tuple}
+    add_types = ATS.parameters
+    rem_types = _try_to_types(RTS)
+    exprs = []
+
+    add_ids = tuple([_component_id(W.parameters[1], T) for T in add_types]...)
+    rem_ids = tuple([_component_id(W.parameters[1], T) for T in rem_types]...)
+
+    push!(exprs, :(index = world._entities[entity._id]))
+    push!(exprs, :(old_archetype = world._archetypes[index.archetype]))
+    push!(
+        exprs,
+        :(
+            new_arch_index =
+                _find_or_create_archetype!(
+                    world, old_archetype.node, $add_ids, $rem_ids,
+                )
+        ),
+    )
+    push!(exprs, :(new_archetype = world._archetypes[new_arch_index]))
+
+    push!(exprs, :(entity_and_row = _create_entity!(world, new_arch_index)))
+    push!(exprs, :(new_entity = entity_and_row[1]))
+    push!(exprs, :(new_row = entity_and_row[2]))
+
+    push!(
+        exprs,
+        :(
+            for comp in old_archetype.components
+                if !_get_bit(new_archetype.mask, comp)
+                    continue
+                end
+                _copy_component_data!(world, comp, index.archetype, new_arch_index, index.row, UInt32(new_row))
+            end
+        ),
+    )
+
+    for i in 1:length(add_types)
+        T = add_types[i]
+        stor_sym = Symbol("stor", i)
+        col_sym = Symbol("col", i)
+        val_expr = :(add.$i)
+
+        push!(exprs, :($stor_sym = _get_storage(world, $T)))
+        push!(exprs, :(@inbounds $col_sym = $stor_sym.data[new_arch_index]))
+        push!(exprs, :(@inbounds $col_sym[new_row] = $val_expr))
+    end
+
+    push!(exprs, :(
+        if _has_observers(world._event_manager, OnCreateEntity)
+            _fire_create_entity(world._event_manager, new_entity, new_archetype.mask)
+        end
+    ))
+
+    push!(exprs, Expr(:return, :new_entity))
+
+    return quote
+        @inbounds begin
+            $(Expr(:block, exprs...))
+        end
+    end
+end
+
 """
     remove_entity!(world::World, entity::Entity)
 
@@ -472,6 +560,78 @@ end
             $(Expr(:block, exprs...))
         end
     end
+end
+
+"""
+    @copy_entity!(world::World, entity::Entity; add::Tuple, remove::Tuple)
+
+Copies an [`Entity`](@ref), optionally adding and/or removing components.
+
+Mutable components are copied by reference, so both entities will share the component instance.
+
+Macro version of [`copy_entity!`](@ref) for more ergonomic component type tuples.
+
+# Example
+
+```jldoctest; setup = :(using Ark; include(string(dirname(pathof(Ark)), "/docs.jl"))), output = false
+entity1 = @copy_entity!(world, entity)
+
+entity2 = @copy_entity!(world, entity;
+    add=(Health(100),),
+    remove=(Position, Velocity),
+)
+
+# output
+
+Entity(0x00000004, 0x00000000)
+```
+"""
+macro copy_entity!(world_expr, entity_expr)
+    :(copy_entity!($(esc(world_expr)), $(esc(entity_expr))))
+end
+macro copy_entity!(kwargs_expr, world_expr, entity_expr)
+    map(x -> (x.args[1] == :remove && (x.args[2] = :(Val.($(x.args[2]))))), kwargs_expr.args)
+    quote
+        copy_entity!(
+            $(esc(world_expr)),
+            $(esc(entity_expr));
+            $(esc.(kwargs_expr.args)...),
+        )
+    end
+end
+
+"""
+    copy_entity!(world::World, entity::Entity; add::Tuple, remove::Tuple)
+
+Copies an [`Entity`](@ref), optionally adding and/or removing components.
+
+Mutable components are copied by reference, so both entities will share the component instance.
+
+For a more convenient tuple syntax, the macro [`@copy_entity!`](@ref) is provided.
+
+# Example
+
+```jldoctest; setup = :(using Ark; include(string(dirname(pathof(Ark)), "/docs.jl"))), output = false
+entity1 = copy_entity!(world, entity)
+
+entity2 = copy_entity!(world, entity;
+    add=(Health(100),),
+    remove=Val.((Position, Velocity)),
+)
+
+# output
+
+Entity(0x00000004, 0x00000000)
+```
+"""
+@inline function copy_entity!(world::World, entity::Entity; add::Tuple=(), remove::Tuple=())
+    if !is_alive(world, entity)
+        throw(ArgumentError("can't copy a dead entity"))
+    end
+    if isempty(add) && isempty(remove)
+        return @inline _copy_entity!(world, entity)
+    end
+    return @inline _copy_entity!(world, entity, Val{typeof(add)}(), add, remove)
 end
 
 """
@@ -978,6 +1138,26 @@ end
         push!(exprs, :(
             if comp == $i
                 _move_component_data!(world._storages.$i, old_arch, new_arch, row)
+            end
+        ))
+    end
+    return Expr(:block, exprs...)
+end
+
+@generated function _copy_component_data!(
+    world::World{CS},
+    comp::UInt8,
+    old_arch::UInt32,
+    new_arch::UInt32,
+    old_row::UInt32,
+    new_row::UInt32,
+) where {CS<:Tuple}
+    n = length(CS.parameters)
+    exprs = Expr[]
+    for i in 1:n
+        push!(exprs, :(
+            if comp == $i
+                _copy_component_data!(world._storages.$i, old_arch, new_arch, old_row, new_row)
             end
         ))
     end