+++ test/runtests.jl
@@ -4,11 +4,18 @@ ENV["ARK_RUNNING_TESTS"] = true
 using Test
 using JET
 
-const N_fake = 0
-const offset_ID = 0
-
 include("include_internals.jl")
+
+if "--large-world" in ARGS
+    include("setup_large.jl")
+else
+    include("setup_default.jl")
+end
+
 include("TestTypes.jl")
+
+include("test_world.jl")
+include("test_query.jl")
 include("test_subarray.jl")
 include("test_structarray.jl")
 include("test_readme.jl")
@@ -17,12 +24,10 @@ include("test_pool.jl")
 include("test_lock.jl")
 include("test_mask.jl")
 include("test_event.jl")
-include("test_query.jl")
 include("test_batch.jl")
 include("test_registry.jl")
 include("test_vec_map.jl")
 include("test_graph.jl")
-include("test_world.jl")
 include("test_quality.jl")
 
 ENV["ARK_RUNNING_TESTS"] = false